refactor: Rename project from secure_share to endless_storage.
