+++ docker-compose.yml
@@ -19,7 +19,7 @@ services:
       - ./server:/app
     env_file:
       - .env.development
-    command: uvicorn secure_share.asgi:application --host 0.0.0.0 --port 8000 --reload
+    command: uvicorn endless_storage.asgi:application --host 0.0.0.0 --port 8000 --reload
     depends_on:
       - db
