+++ server/.gitignore
@@ -135,4 +135,4 @@ DS_Store
 # Temporary Files.
 temp
 
-# secure_share
+# endless_storage
