+++ server/accounts/models.py
@@ -2,7 +2,7 @@
 from django.contrib.auth.models import AbstractBaseUser
 from django.db import models
 
-from secure_share.models import BaseModel
+from endless_storage.models import BaseModel
 
 from .managers import UserManager
