+++ server/drive/tests/test_drive_members.py
@@ -2,7 +2,7 @@ from django.contrib.auth import get_user_model
 from django.urls import reverse
 from rest_framework import status
 
-from secure_share.tests import BaseTestCase
+from endless_storage.tests import BaseTestCase
 
 from ..constants import DriveMemberRole
 from ..models import DriveMember
