+++ server/endless_storage/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for secure_share project.
+ASGI config for endless_storage project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "secure_share.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "endless_storage.settings")
 
 application = get_asgi_application()
