+++ server/endless_storage/exceptions/handler.py
