+++ server/endless_storage/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for secure_share project.
+Django settings for endless_storage project.
 
 Generated by 'django-admin startproject' using Django 5.1.4.
 
@@ -73,7 +73,7 @@ MIDDLEWARE = [
     "corsheaders.middleware.CorsMiddleware",
 ]
 
-ROOT_URLCONF = "secure_share.urls"
+ROOT_URLCONF = "endless_storage.urls"
 
 TEMPLATES = [
     {
@@ -91,7 +91,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "secure_share.wsgi.application"
+WSGI_APPLICATION = "endless_storage.wsgi.application"
 
 
 # Database
@@ -180,8 +180,8 @@ TRENCH_AUTH = {
 
 
 REST_FRAMEWORK = {
-    "EXCEPTION_HANDLER": "secure_share.exceptions.handler.custom_exception_handler",
-    "DEFAULT_RENDERER_CLASSES": ["secure_share.response_renderer.CustomJsonRender"],
+    "EXCEPTION_HANDLER": "endless_storage.exceptions.handler.custom_exception_handler",
+    "DEFAULT_RENDERER_CLASSES": ["endless_storage.response_renderer.CustomJsonRender"],
     "DEFAULT_AUTHENTICATION_CLASSES": (
         "rest_framework_simplejwt.authentication.JWTAuthentication",
     ),
