+++ server/endless_storage/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for secure_share project.
+WSGI config for endless_storage project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "secure_share.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "endless_storage.settings")
 
 application = get_wsgi_application()
